Don't localize $a for List::MoreUtils::pairwise
This was causing bizarre "Attempt to free unreferenced scalar" error messages
from Perl, and replacing the last value in @$observed with something false.
(Possibly the underlying problem is a bug in pairwise().)

Further, we get exactly the same value returned if we calculate the relevant
sum manually (rather than using pairwise()), so I'm reasonably confident
that this doesn't introduce any new bugs.